test: fix leakage between tests
When mocking out `homedir` we were leaking data between tests.
This results in tests that were passing when they shouldn't and vice versa.

Now we always mock `homedir` and do so in a way that is contained to the test being run.

Affected tests have been cleaned up as necessary.