feat[pages]: Add new internal pages functions `optimize-routes` command (#1648)
This PR adds a new pages functions command called `optimize-routes`. This is an
internal-use ONLY command at the moment, and is therefore not exposed to external
consumers. The command is meant to parse a given  `_routes.json` file, optimize
and consolidate the route paths defined in it, and output the resulting optimized
file.

Co-authored-by: Carmen Popoviciu <cpopoviciu@cloudflare.com>