Allow `--experimental-local` to be used with module workers
Previously, Miniflare 3 was automatically collecting modules.
These collected modules were given names relative to the current
working directory. Because Wrangler puts scripts in a temporary
directory, these names started with `..`, which `workerd` was not
happy with.

This change avoids the redundant automatic collection in Miniflare,
and ensures module names are resolved relative to the temporary
directory.