Fix importing npm packages with same name of Node built-ins
Previously, if `node_compat` was disabled, and the user had an npm package with
the same name as a Node built-in module installed (e.g. `buffer`), esbuild
would always throw instead of resolving it. This PR fixes the issue by relying
on esbuild's resolver, and instead rewriting the error message on failure to
suggest enabling `node_compat` instead of `platform: "node"`.

Closes #2038