Support remote KV storage in `--experimental-local`
Adds a new `--experimental-local-remote-kv` flag that adds
experimental support for reading/writing from/to real KV namespaces.
Internally, the option is refered to as `experimentalLocalRemoteKv`
(not `...KV`) since that's what Yargs converts it to, and I think
it's better to be consistent with other options here.

This also upgrades `@miniflare/tre` to `3.0.0-next.6` because I
forgot to include the accessing private networks fix in `.5`.