Fix file-watching when using middleware with service workers
Previously, when middleware was enabled, service worker user code was
being bundled as part of the middleware facade application, not in
the final bundling step with `watch` enabled. This meant changes to
user code were not picked up. This change restructures the service
worker middleware loader and bundling code as an IIFE, that gets
injected into the final bundle. This also has the positive side
effect that middleware internals aren't exposed to users.