Log uncaught source-mapped errors to the console (#446)
Previously, errors _thrown_ using the `MF-Experimental-Error-Stack`
header were only shown in a pretty-error page. This PR logs those
errors with source-maps applied to the console too, using the
`source-map-support` package. This simplifies our code, and also
means we don't need to touch internal Youch methods, as the errors
we pass to Youch are already source-mapped.