fix: fallback to copy for failed renames when generating projects
The internals of the generate subcommand clone a repo (usually the
cloudflare/templates one) into a temp directory and then move the
relevant template to the generated project, but on Linux a rename
operation might not always work across devices. This meant that any
Linux user trying to generate a Worker from a template would get an
error no matter what. To fix this Wrangler now falls back to copying the
template to the destination and then deleting the copied directory.