Support Queue events in tail
Queue events have an "event" section containing two fields -- "queue"
and "batchSize", which contain the queue name and number of messages in
the batch. Queue events can be distinguished from other events by the
presence of the "queue" field, so that's what I do here.

I've followed the example of RequestEvent more closely than either
ScheduledEvent or AlarmEvent here, but am happy to change up the format
if there's a reason to do so.