Allow `--experimental-local` to cleanly exit on `x`/`CTRL-C` (#2400)
* Allow `--experimental-local` to cleanly exit on `x`/`CTRL-C`

In the bootstrapper script `bin/wrangler.js`, we open an IPC channel,
so IPC messages from the Wrangler process are propagated through the
bootstrapper. Normally, Node's SIGINT handler would close this for
us, but interactive mode enables raw mode on stdin which disables the
built-in handler. This PR calls `process.disconnect()` to close this
channel when we press `x` or CTRL-C, ensuring no open handles, and
allowing for a clean exit.

Closes #2387

* fixup! Allow `--experimental-local` to cleanly exit on `x`/`CTRL-C`

* fixup! fixup! Allow `--experimental-local` to cleanly exit on `x`/`CTRL-C`

* fixup! Allow `--experimental-local` to cleanly exit on `x`/`CTRL-C`