docs(templates): add a service binding example to new-worker.ts
with the release of service bindings, it's helpful to have an immediate idea of how service bindings work.  especially in typescript land, where the type is useful info.  in my case, i found `Fetcher` to be the most accurate type, but feel free to correct me if this is invalid.