Disable format conversion when detecting format, closes #1668
Wrangler automatically detects whether your code is a `modules` or
`service-worker` format Worker based on the presence of a `default`
`export`. This check currently works by building your entrypoint with
`esbuild` and looking at the output metafile.

Previously, we were passing `format: "esm"` to `esbuild` when
performing this check, which enables *format conversion*. This may
introduce `export default` into the built code, even if it wasn't
there to start with, resulting in incorrect format detections.

This change removes `format: "esm"` which disables format conversion
when bundling is disabled: https://esbuild.github.io/api/#format.

We may want to use a package like `es-module-lexer` in the future,
but this issue is affecting users, and this is probably the simplest
fix.

Closes #1668
Closes #2396
Ref #2737