[Miniflare 3] Add support for D1 (#519)
* Log error `cause`s and revive errors with correct prototype

Previously, we were ignoring `cause`s on `Error`s. These are used a
lot by D1, so in preparation for supporting that, allow a `cause`
property to be set on `JsonError`s passed with the
`MF-Experimental-Error-Stack` header. This will need a change in
`wrangler`'s `middleware-miniflare3-json-error.ts` file.

This change also attempts to use the correct error constructor (e.g.
`TypeError`) when reviving errors from JSON based on the `name`. If
the correct class can't be found, we just fall back to `Error`. This
will be useful when propagating errors through `dispatchFetch()`.

* Propagate uncaught errors through `dispatchFetch()`

If we receive a `JsonError` with a `MF-Experimental-Error-Stack`
header in `dispatchFetch()` responses, we now rethrow the revived
error. This matches the behaviour of Miniflare 2, and makes testing
thrown errors easier.

Note this has the same caveats as `MF-Experimental-Error-Stack`, in
that it requires code modifications to work. We're unable to catch
errors from service binding calls yet.

* Add support for D1

This change ports Miniflare 2's D1 implementation to Miniflare 3.
We'd already laid lots of the groundwork for this by reworking
Miniflare 2 to implement D1's underlying API instead of the Worker
binding API.

Miniflare 3 will add a regular `Fetcher` service binding for each
database specified in `d1Databases`. Note this is different to
Miniflare 2 which provided actual `D1Database` bindings. If we
wanted to replicate this behaviour, we'd need to transform code
similar to what Wrangler does at the moment. This needs a small
change to the D1 facade application code, as `--experimental-local`
now needs to be treated like remote mode. The facade itself also
needs to be updated to pass full URLs to the underlying service
bindings, as `workerd` does not provide host-less `Fetcher`s.

* fixup! Log error `cause`s and revive errors with correct prototype

* fixup! Add support for D1