Reuse `SourceMapConsumer` between source-mapping operations
Previously, we were reading and parsing source-maps every time we
received an error from a Worker to log. Parsing source-maps can be
quite expensive, so this change lifts this out into a memoised
`Promise`, that gets reset whenever the script changes. Note, when a
user sees an error in dev, it's quick likely they'll edit their
script to fix it, so this isn't too much of a problem at the moment.
We're going to add `cause` support now though, which may require
source-mapping multiple errors in quick succession.