Log `DOMException` causes in `dev`
`DOMException`s are special, the actually useful stack trace is
hidden behind the `stack` getter, and not included in the preview.
`DOMException`s are quite common (e.g. invalid base64 in `atob()`,
invalid algorithm name when using WebCrypto, ...). Currently, these
errors are just logged as `DOMException`, which isn't helpful for
debugging. With this change, the actual reason is logged, along with
the source-mapped location.