ci: fix the release github workflow
This workflow was failing whenever there was a minor release to wrangler
and a patch release to a package that depended upon wrangler.
It was caused by a bug in npm that was resolved in v8.5.0.
See https://github.com/npm/cli/issues/4379 and https://github.com/npm/cli/pull/4371.
The error looks like: `No matching version found for wrangler@^2.15.0.`