Bump to TypeScript 5 and fix bundled TypeScript output (#635)
* Bump to `typescript@5.0.4` and `@microsoft/api-extractor@7.63.3`

TypeScript 5 includes lots of goodies, and allows us to remove a
bunch of `@ts-expect-error`s.

* Set `"moduleResolution": "bundler"` in `tsconfig.json`

This loosens some resolution rules in accordance with how bundlers
like `esbuild` resolve modules. This also allows us to remove a bunch
of `@ts-expect-error`s.

* Remove import cycle in `src/plugins/r2`

This probably wasn't a problem, but generally it's not great to have
cyclic dependencies. For future reference, this was found with
`npx madge --circular --extensions ts packages/miniflare/src`.