Implement generic logging endpoint in loopback server
This replaces the loopback server's support for logging requests with
a more generic endpoint, that supports logging any arbitrary message.
This will be needed by Durable Object simulators that want to write
to the console, without `workerd`'s verbose prefixing. In particular,
Queues logs dispatches, and Cache logs first usage when the
`cacheWarnUsage` option is enabled.