Allow workers to be accessed directly via dedicated sockets
We'd like this for Wrangler, so we can use the same Miniflare
instance for both the regular proxy worker and the inspector proxy
worker. The inspector proxy worker needs to run on its own port as
`/json` and `/json/version` are well-known path names used by
`chrome://inspect`'s service discovery. Otherwise, we'd be able to
use `routes`. Note this is an unsafe option as it bypasses
Miniflare's entry worker, so won't include any request logging/
pretty errors.