Remote TurboRepo CI/CD
Setting up TurboRepo Custom Remote Cache to now work in the CI/CD GitHub Workflows. Removed the 'remote:' scripts and added the .env injection to the
original scripts, allowing for anyone or CI to use the original scripts to utilize TurboRepo Remote Cache, if there is no .env it falls back to local mode.