Warn that Hyperdrive bindings do not yet work in local dev
Support is already in progress (see
https://github.com/cloudflare/workerd/pull/1266) and should be coming in
the next few weeks, but we've already had at least one user get very
confused by this and cause us to spend time on debugging before
realizing the problem was that they were talking about problems in
wrangler dev rather than at the edge. This is an easy way to attempt to
mitigate that for now.