Only show compatibility date fallback warning if new version available
When a compatibility date greater than the installed `workerd` version
is selected, Miniflare logs a warning. The intention here is to warn
users if they've set a compatibility date, but haven't updated
`workerd`, so features they've requested won't be available.
The problem is that this warning is only actionable if there's an
update available. This change suppresses the warning if there's no
new version available.