Use installed `workerd` version as default compat date instead of now
`wrangler dev` would previously use the current date as the default
compatibility date if one wasn't set in `wrangler.toml`. Before the
previous commit, this would show a warning if it was greater than the
installed `workerd` version. Whilst using the current date as the
default would no longer log a warning with this change, if there was
an update it would. This change sets the default to the `workerd`
version instead.