feat: add support for wrapped bindings
This change adds a new `wrappedBindings` worker option for configuring
`workerd`'s wrapped bindings. These allow custom bindings to be
written as JavaScript functions accepting an `env` parameter of
"inner bindings" and returning the value to bind.

The `wrappedBindings` option maps binding names to designators.
The worker defined by the designator is used for the wrapped binding's
source and inner bindings. The worker becomes un-routable, and cannot
be used for service or Durable Object bindings. Using a worker allows
us to piggyback on all of Miniflare's existing bindings. For example,
the wrapped binding worker can declare a function-valued service
binding for accessing Node.js state. JSON bindings can be specified
with the designator to override specific bindings.