ci: correctly update the lockfile when generating the Version PR
Previously the changeset-version.js script was failing in CI because we
were using `pnpm i`, which (in CI) defaults to a frozen lockfile.
This then failed, since we were actually trying to update the lockfile
with this command.

Now we explicitly tell pnpm to only update the lockfile and not worry about
actually installing anything.