[miniflare] Add support for "sticky blobs" that aren't deleted
For isolated per-test storage, we'll be pushing/popping databases
to a "stack". After pushing a database stack frame, we may delete a
record in a test. We'd like to avoid deleting the associated blob here
so when we pop the frame, the record still links up with it.

This change adds an `unsafeStickyBlobs` option that disables garbage
blob cleanup if enabled.