[miniflare] Add support for colo-local ephemeral "Durable" Objects
For isolated storage, we'd like the runner object not to persist any
state to disk, so we don't have to worry about excluding it when
managing the storage stack. We currently have an
`unsafeEphemeralDurableObjects` option that makes _all_ Durable
Objects in a worker use in-memory state. This isn't suitable for our
Vitest pool, as user objects will run in the same worker as the runner
object.

This change adds support for `workerd` "colo-local" actors. These are
like Durable Objects, but don't provide any durable storage (i.e. no
`state.storage` API), and have a slightly different namespace binding
API.