[miniflare] Bind `this` to `Miniflare` instance in custom services
We'd like to extend our Vitest pool's custom loopback service with
endpoints for managing isolated storage stacks. To do this, we need to
know the persistence paths of the `Miniflare` instance we're currently
running in. A previous commit added an `unsafeGetPersistPaths()`
function. If we wanted to access these paths in the pool's loopback
service, we'd need to capture the paths/`Miniflare` instance in a
closure used as the service binding function. Unfortunately, we use
`util.deepStrictEqual()` on Miniflare options to determine if we
should restart the `Miniflare` instance. This would always return
`true` if we created a new bound/arrow function on each test run.

This change binds `this` to the current `Miniflare` instance when
calling custom service functions, meaning we can reuse the same
function when building options.