[miniflare] Add support for `rootPath`s
Previously, Miniflare would always resolve path-valued options
relative to the current working directory. For the Vitest pool, we'd
like to be able to resolve these relative to the project config file.

This change adds back Miniflare 2's `rootPath` option for controlling
the resolution directory. `rootPath` can be set as a shared or
per-worker option. Path-valued options resolve relative to the closest
`rootPath`.