[C3] Ensure `ts` is defined (#4598)
* Ensure ts is defined

* Add e2e testing for worker templates

* Prevent hard-wrapping with `log-update`

Previously log-update was hard-wrapping at 80 chars in e2e tests, which broke regex matching on output data. This is because process.stdout.columns isn't inherited in child processes, and 80 is `log-update`'s default. Since hard-wrapping is not what we want anyway, this modifies `@cloudflare/cli` to pass in a `process.stdout` object with `Number.MAX_SAFE_INTEGER` columns, which effectively disables hard wrapping.

* Not so wide

* Remove npm svelte

* Skip e2e workers tests on Windows