ci: only deploy Workers when pushing to special branches (#4848)
Previously every push to `main` that contained changes to a Worker package would cause that Worker to be deploy "to production".
This meant that we had to be very careful about merging PRs that touch these Workers files.

Now production deployment is only run for a Worker on a push to its special deployment branch (e.g. `deploy-worker/playground-preview-worker`).
This allows us to land PRs to `main` and then collect up changes to release in a single go by updating the special deployment branch in a controlled way.
Moreover it is also easy to see what changes will be released by looking at the difference between `main` and the deployment branch for the Worker package directory.

The three workers in this repo now follow this pattern:

- edge-preview-authenticated-proxy
- format-errors
- playground-preview-worker

In addition, the Playground Preview Worker has end-to-end tests that are run on every push to the `main` branch, and also any PRs that have the "playground-worker" label.
For these tests the Worker is deployed to the testing environment before running the tests against this deployment.