Partial refactor: remove zoneID from `getZoneIdHostAndRoutes()`
`zoneId` is only fetched in remote mode. Whether or not Wrangler is in remote mode can be decided further down the call stack than this function, and so when Wrangler is switched between local mode and remote mode the `zoneId` value returned from this function is stale. The `host` and `routes` returned from this function are always the same regardless of remote mode/local mode