ci: disable file parallelism for fixture tests
Some of the `get-{bindings,platform}-proxy` tests have been flaking in
CI recently. I think this is because they're trying to connect to the
dev registry, but the dev registry was started by an earlier test that
just finished. This leads to socket errors like `other side closed`.
The proper solution to this problem is to start a dev registry for
each test (file). We could do this by assigning a different dev
registry port for each test (file/project). In the meantime, this
change disables parallelism for the fixture tests, ensuring if a dev
registry is started, it is definitely shut down before another test
begins.