feat(wrangler): Implement environment inheritance for Pages
To make it easier for users to write config files for Pages,
Wrangler will not require both of those environments to be
explicitly defined in the `wrangler.toml` file. If either
`[env.production]` or `[env.preview]` is left unspecified,
Wrangler will use the top-level Wrangler environment when
targeting that named Pages environment.

This commit implements the aforementioned environment
inheritance logic for Pages projects.