chore: ensure that dependabot updates package-lock.json
Previously we limited the workerd updates to the miniflare package by setting the `directory` configuration.
But it seems that this fails to update the package lockfile because that is higher up in the root of the repo.

Changing the `directory` back to `/` should cover the package lockfile, although it will also update all packages and fixtures that contain potential updates (notably for workers-types).
But this is probably a good thing overall.