Set `Accept-Encoding: identity` for AI API requests
Streaming doesn't currently work with `wrangler dev` if the responses are gzipped. This may be due to workerd not being able to stream gzipped bodies – still under investigation. But in the meantime we can explicitly opt out of gzip responses by setting `Accept-Encoding: identity` when making requests to the API.