Fix the worked-speedtest template (#5830)
The code had been converted to typescript in 026d944d, but packages.json and several other places still refered to .js files.

In addition a refactoring of the test code meant tests for both /up and /down urls were now being sent to the handler for /up, causing the tests to fail. I removed some of the test in up.test.ts since they were testing different values of the ?bytes= parameter, but this parameter is only used in down.ts, not in up.ts.