chore: ensure playground-preview-worker uses the workspace dependency of wrangler
Without this change the changesets tooling breaks when trying to update both this
package and wrangler at the same time.

The error that can be seen was:

```
ERR_PNPM_NO_MATCHING_VERSION No matching version found for wrangler@^3.57.0

This error happened while installing a direct dependency of /home/runner/work/workers-sdk/workers-sdk/packages/playground-preview-worker

The latest release of wrangler is "3.56.0".
```